Reworked line printing
An attempt to handle issue #1

The older method of printing strings could not accomodate
characters that had a visual length other than 1 column
Which caused visual desync of the terminal cursor position
and the the line being edited.

This commit attempts to solve this problem by having the
print_line function attempt to shift the position of the
terminal cursor.

There might be certain quirks to this method which might become
nuiscance if not handeled properly.

[!] Further development needed